[DDW-321] Separates Addresses requests into their own sub-directory
