+++ source/renderer/app/api/addresses/requests/getAddress.js
@@ -1,5 +1,6 @@
 // @flow
-import type { AdaAddress, RequestConfig } from './types';
+import type { RequestConfig } from '../../common/types';
+import type { Address } from '../types';
 import { request } from '../../utils/request';
 
 export type GetAddressParams = {
@@ -9,7 +10,7 @@ export type GetAddressParams = {
 export const getAddress = (
   config: RequestConfig,
   { address }: GetAddressParams
-): Promise<AdaAddress> => {
+): Promise<Address> => {
   const encodedAddress = encodeURIComponent(address);
   return request({
     hostname: 'localhost',
